+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - File name changed to catherinefbhd.suprx
 - OSD indicating patch success or failure
+- Scale title logos